pyproject.toml
[project]
name = "parking-tracker"
version = "0.1.0"
authors = [
    { name = "Jassiel Ovando", email = "a2021117600@estudiantes.upsa.edu.bo" },
    { name = "Mateo Fernandez" },
    { name = "Emiliana Ribera" },
]
description = "Parking tracker app for android auto."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "markdown>=3.8",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "ruff>=0.11.9",
    "ultralytics>=8.3.144",
]
